Add initial support for Nextcloud 34
Nextcloud 34 removed the private OC\Files\Mount\MoveableMount interface
in favour of the public OCP\Files\Mount\IMovableMount (server commit
fbf84e64732, "chore: Replace MoveableMount usage with IMovableMount"),
so the anonymous mount-point class in MountProvider failed to
instantiate on Nextcloud 34.

Extract that anonymous class into the named class ArchiveMountPoint,
implementing the public IMovableMount interface, which is all Nextcloud
>= 34 checks for. Since Nextcloud <= 33 marks a mount point as
movable/removable only through the private MoveableMount interface, add
a LegacyArchiveMountPoint subclass which additionally implements it;
MountProvider picks the legacy variant at runtime when that interface
still exists, keeping move and delete of archive mount points working on
Nextcloud 33.

All other APIs used by the app (public and private) were checked against
nextcloud/server stable34: no other class, method or relevant signature
used by the app was removed or changed incompatibly.

Bump the supported Nextcloud version range to 33-34 accordingly.

Tested against Nextcloud 34.0.2 RC1 (and re-tested on 33.0.6): mounting,
browsing, reading, moving and deleting archive mounts, archive
extraction, admin/personal settings, MIME type repair step, and the file
actions in the web UI all work.
