+++ appinfo/info.xml
@@ -12,7 +12,7 @@
   <category>files</category>
   <bugs>https://github.com/rotdrop/nextcloud-app-files-archive.git</bugs>
   <dependencies>
-    <nextcloud min-version="33" max-version="33"/>
+    <nextcloud min-version="33" max-version="34"/>
     <php min-version="8.4" max-version="8.5" />
   </dependencies>
   <types>
